Switch from tensor-form to (partial) stress form viscosity.
This is the correct form for LES. It should make little or no difference for
isoviscous cases. I think we should simply use it for all cases.

In test_multi_steady_state.py: don't overwrite other defaults when setting a
tight tolerance for the newton solver. Same 3 tests failing as before:
- gradient_verification_configurations
- optimal_friction_mini_model
- partial_derivative_turbine