Add new "free_slip" boundary condition.
This is to replace the way we applied free slip before: setting a
weak_dirichlet with a Constant((0.0,0.0)) as value. Because we don't actually
enforce the tangential components weakly, this actually produced a free slip
rather than a no-slip condition. However this approach did add in an extra
boundary term associated with the advection term, which kicks in for
u.n<0 (at the inflow). Even though this extra term is zero when u.n is
exactly zero, its nonlinear behaviour around 0 causes issues in
gradient calculations. The new "free_slip" bctype simply leaves this term out entirely.

Replaced all the cases in the tests and examples where free slip was enforced
via "weak_dirichlet" with the new "free_slip". There are still some tests that test
"weak_dirichlet" for nonzero bc expressions. NOTE that the remaining "weak_dirichlet"
applies a Dirichlet on both components for advection-only cases only
(or negligible viscosity). With no advection and no viscosity, it applies a normal condition
only on the continuity equation. In other cases, e.g. advection and considerable viscosity,
or viscosity and no advection, it is probably inconsistent.