Implement correction to discrete turbine parameterisation.
This fixes both a bug in the integral of the bum function (which should
have had been scaled with D**2 instead of D), and the correction from:

SC Kramer, MD Piggott, A correction to the enhanced bottom drag
parameterisation of tidal turbines, Renewable Energy 92, 2016
http://doi.org/10.1016/j.renene.2016.02.022

This is all done in the existing BumpTurbine which is now initialized
by specifying a thrust_coefficient instead of the `friction` scalar.
TODO: implement thrust table.

Removes ThrustTurbine (which partly implemented the correction but also
had bugs) and ImplicitThrustTurbine (which was never developed). Cut in
and rated speed functionality to be reimplemented via thrust curve
table.

Passes all tests except those in tests/turbine_correction which uses
ThrustTurbine. TODO: change those to use BumpTurbine