Reimplement thrust curve
Thrust curve for BumpTurbine now consistent with power and force
calculation specified as a list of (u,Ct) pairs.
Add helper function standard_thrust_curve() that produces a standard
thrust curve with cut-in, rated,and cut-out speeds.
To simplify things depth is now be provided at the creation of the
BumpTurbine rather than when force and power are calculated.
Clean up sphinx documentation.
Add test_thrust_curve that tests whether the applied force and power
in the hydrodynamic model agrees with the analytical actuator disc
results.
All tests are passing.